// src/app/shared/Models/serviceRequestObj.ts
export class serviceRequestObj {
    serviceRequestID: any;
    serviceRequestCutomerSummary: any;
    serviceRequestEmployeeSummary: any;
    serviceRequestDate: Date;
    assignmentToEmpDate: Date;
    userId: any;
    employeeId: any;
    voiceNotePath: any;
    serviceId: any;
    customerPackageId: any;
    isDeleted: any;
    orderType: any;
    creationDate: Date;
    orderTypeName: any;
    orderTypeNameAr: any;
}